pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"

dependencies = [
  "flask",
  "python-dotenv",
  "gunicorn",
  "psycopg2",
  "validators"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

